+++ __test__/create.test.js
@@ -36,66 +36,113 @@ describe('Create', function () {
 
     $(document.body).append(listEl)
 
-  describe('Without items and with template function', function() {
-    var listEl = $('<div id="list">\
-      <ul class="list"></ul>\
-    </div>');
+    var list = new List(
+      'list',
+      {
+        valueNames: ['name'],
+        item: '<li><span class="name"></span></li>',
+      },
+      [{ name: 'Jonny' }]
+    )
 
-    $(document.body).append(listEl);
+    it('should contain one item', function () {
+      expect(list.items.length).toEqual(1)
+      expect(listEl.find('li').length).toEqual(1)
+    })
 
-    var list = new List('list', {
-      valueNames: ['name'],
-      item: function(values) {
-        return `<li data-template-fn-${values.name.toLowerCase()}><span class="name"></span></li>`;
-      }
-    }, [
-      { name: 'Jonny' }
-    ]);
+    it('should contain two items', function () {
+      list.add({ name: 'Jonas' })
+      expect(list.items.length).toEqual(2)
+      expect(listEl.find('li').length).toEqual(2)
+    })
 
-    it('should contain one item', function() {
-      expect(list.items.length).toEqual(1);
-      expect(listEl.find('li').size()).toEqual(1);
-    });
+    listEl.remove()
+  })
 
-    it('should contain two items', function() {
-      list.add({ name: 'Jonas' });
-      expect(list.items.length).toEqual(2);
-      expect(listEl.find('li').size()).toEqual(2);
-    });
-
-    it('should get values from items', function() {
-      list.add({ name: 'Egon' });
-      expect(listEl.find('li[data-template-fn-egon]').size()).toEqual(1);
-    });
-
-    listEl.remove();
-  });
-
-  describe('Without items and with string template for table', function() {
+  describe('Without items and with string template for table', function () {
     var listEl = $('<div id="list">\
       <table class="list"></table>\
     </div>')
 
     $(document.body).append(listEl)
 
-  describe('without items and or template', function() {
+    var list = new List(
+      'list',
+      {
+        valueNames: ['name'],
+        item: '<tr><span class="name"></span></tr>',
+      },
+      [{ name: 'Jonny' }]
+    )
 
-    it('should throw error on init', function() {
+    it('should contain one item', function () {
+      expect(list.items.length).toEqual(1)
+      expect(listEl.find('tr').length).toEqual(1)
+    })
+
+    it('should contain two items', function () {
+      list.add({ name: 'Jonas' })
+      expect(list.items.length).toEqual(2)
+      expect(listEl.find('tr').length).toEqual(2)
+    })
+
+    listEl.remove()
+  })
+
+  describe('Without items and with template function', function () {
+    var listEl = $('<div id="list">\
+      <ul class="list"></ul>\
+    </div>')
+
+    $(document.body).append(listEl)
+
+    var list = new List(
+      'list',
+      {
+        valueNames: ['name'],
+        item: function (values) {
+          return `<li data-template-fn-${values.name.toLowerCase()}><span class="name"></span></li>`
+        },
+      },
+      [{ name: 'Jonny' }]
+    )
+
+    it('should contain one item', function () {
+      expect(list.items.length).toEqual(1)
+      expect(listEl.find('li').length).toEqual(1)
+    })
+
+    it('should contain two items', function () {
+      list.add({ name: 'Jonas' })
+      expect(list.items.length).toEqual(2)
+      expect(listEl.find('li').length).toEqual(2)
+    })
+
+    it('should get values from items', function () {
+      list.add({ name: 'Egon' })
+      expect(listEl.find('li[data-template-fn-egon]').length).toEqual(1)
+    })
+
+    listEl.remove()
+  })
+
+  describe('without items and or template', function () {
+    it('should throw error on init', function () {
       var listEl = $('<div id="list">\
         <ul class="list"></ul>\
       </div>')
       $(document.body).append(listEl)
 
-      expect(function() {
+      expect(function () {
         var list = new List('list', {
-          valueNames: ['name']
-        });
-      }).toThrow();
-      listEl.remove();
-    });
-  });
+          valueNames: ['name'],
+        })
+      }).toThrow()
+      listEl.remove()
+    })
+  })
 
-  describe('Without items and with HTML template', function() {
+  describe('Without items and with HTML template', function () {
     var listEl = $('<div id="list">\
       <ul class="list"></ul>\
     </div>')